requirements.txt
llama-cpp-python>=0.2.90
gradio>=4.29.0
datasets>=2.18.0